backend/requirements.txt
amqp>=5.0.9
asgiref>=3.5.0
backports.zoneinfo>=0.2.1
bcrypt>=3.2.0
billiard>=3.6.4.0
black>=22.1.0
boto3>=1.20.47
botocore>=1.23.47
certifi>=2021.10.8
cffi>=1.15.0
charset-normalizer>=2.0.11
click>=8.0.3
click-didyoumean>=0.3.0
click-plugins>=1.1.1
click-repl>=0.2.0
cryptography>=36.0.1
decorator>=5.1.1
Django>=4.0.2
django-cors-headers>=3.11.0
django-rest-framework>=0.1.0
django-rest-knox>=4.2.0
django-vote>=2.2.0
djangorestframework>=3.13.1
dynaconf>=3.1.7
hvac>=0.11.2
idna>=3.3
Jinja2>=3.0.3
jmespath>=0.10.0
kombu>=5.2.3
loguru>=0.6.0
MarkupSafe>=2.0.1
mypy-extensions>=0.4.3
pathspec>=0.9.0
platformdirs>=2.4.1
prompt-toolkit>=3.0.26
psycopg2>=2.9.3
pycparser>=2.21
PyNaCl>=1.5.0
python-dateutil>=2.8.2
pytz>=2021.3
regex>=2022.1.18
requests>=2.27.1
s3transfer>=0.5.1
shortuuid>=1.0.8
six>=1.16.0
sqlparse>=0.4.2
tomli>=2.0.0
typing_extensions>=4.0.1
urllib3>=1.26.8
validators>=0.18.2
vine>=5.0.0
wcwidth>=0.2.5
